Rename 'registerdomain' package into 'domainregister' for consitency and clarity
