+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -113,7 +113,7 @@ import org.wordpress.android.ui.reader.views.ReaderLikingUsersView;
 import org.wordpress.android.ui.reader.views.ReaderSiteHeaderView;
 import org.wordpress.android.ui.reader.views.ReaderTagHeaderView;
 import org.wordpress.android.ui.reader.views.ReaderWebView;
-import org.wordpress.android.ui.registerdomain.suggestionslist.DomainSuggestionsFragment;
+import org.wordpress.android.ui.domainregister.suggestionslist.DomainSuggestionsFragment;
 import org.wordpress.android.ui.stats.StatsAbstractFragment;
 import org.wordpress.android.ui.stats.StatsActivity;
 import org.wordpress.android.ui.stats.StatsConnectJetpackActivity;
