+++ WordPress/src/main/java/org/wordpress/android/ui/ActivityLauncher.java
@@ -52,7 +52,7 @@ import org.wordpress.android.ui.prefs.MyProfileActivity;
 import org.wordpress.android.ui.prefs.notifications.NotificationsSettingsActivity;
 import org.wordpress.android.ui.publicize.PublicizeListActivity;
 import org.wordpress.android.ui.reader.ReaderPostPagerActivity;
-import org.wordpress.android.ui.registerdomain.suggestionslist.DomainSuggestionsActivity;
+import org.wordpress.android.ui.domainregister.suggestionslist.DomainSuggestionsActivity;
 import org.wordpress.android.ui.stats.StatsActivity;
 import org.wordpress.android.ui.stats.StatsConnectJetpackActivity;
 import org.wordpress.android.ui.stats.StatsConstants;
@@ -575,7 +575,7 @@ public class ActivityLauncher {
         }
     }
 
-    public static void viewRegisterDomainActivity(Activity activity, SiteModel site) {
+    public static void viewDomainSuggestionsActivity(Activity activity, SiteModel site) {
         Intent intent = new Intent(activity, DomainSuggestionsActivity.class);
         intent.putExtra(WordPress.SITE, site);
         activity.startActivity(intent);
