+++ WordPress/src/main/res/layout/activity_domain_suggestions_activity.xml
@@ -11,7 +11,7 @@
 
     <fragment
         android:id="@+id/fragment_container"
-        android:name="org.wordpress.android.ui.registerdomain.suggestionslist.DomainSuggestionsFragment"
+        android:name="org.wordpress.android.ui.domainregister.suggestionslist.DomainSuggestionsFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 </LinearLayout>
