+++ WordPress/src/main/res/values/strings.xml
@@ -774,7 +774,7 @@
     <string name="activity_log_rewind_site">Rewind Site</string>
     <string name="activity_log_rewind_dialog_message">Are you sure you want to rewind your site back to %1$s at %2$s? This will remove all content and options created or changed since then.</string>
 
-    <!-- Register Domain -->
+    <!-- Domain Registration-->
     <string name="register_domain">Register Domain</string>
     <string name="choose_domain">Choose Domain</string>
     <string name="suggestions_empty_list">No suggestions found</string>
